# docs/swagger.yaml
basePath: /api/v1
definitions:
  payments.CheckoutSessionRequest:
    properties:
      amount:
        type: string
    required:
    - amount
    type: object
  payments.RefundRequest:
    properties:
      transaction_ref:
        type: string
    required:
    - transaction_ref
    type: object
  users.CreateUserTokenRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  users.RegisterMerchantRequest:
    properties:
      business_name:
        type: string
      city:
        type: string
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      post_code:
        type: string
      state:
        type: string
      street:
        type: string
    required:
    - business_name
    - email
    - first_name
    - last_name
    - password
    type: object
  users.RegisterRequest:
    properties:
      city:
        type: string
      dob:
        type: string
      email:
        type: string
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      password:
        type: string
      phone_number:
        type: string
      post_code:
        type: string
      state:
        type: string
      street:
        type: string
    required:
    - email
    - first_name
    - last_name
    - password
    type: object
  wallets.PayoutWebhookPayload:
    properties:
      amount:
        type: string
      event_id:
        type: string
      gateway_ref:
        type: string
      source_id:
        description: for our PSP 645b44fb-314f-4e34-8106-17b09cc9660a
        type: string
      status:
        type: string
    type: object
  wallets.TopupRequest:
    properties:
      amount:
        type: string
    required:
    - amount
    type: object
  wallets.TopupWebhookPayload:
    properties:
      amount:
        type: string
      event_id:
        type: string
      gateway_ref:
        type: string
      source_id:
        description: for our PSP 645b44fb-314f-4e34-8106-17b09cc9660a
        type: string
      status:
        type: string
    type: object
  wallets.TransferRequest:
    properties:
      amount:
        type: string
      email:
        type: string
      reference:
        type: string
    required:
    - amount
    - email
    - reference
    type: object
  wallets.WithdrawalRequest:
    properties:
      amount:
        type: string
      bank_name:
        type: string
      number:
        type: string
      reference:
        type: string
    required:
    - amount
    - reference
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is digital wallet API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Wallet API
paths:
  /auth/register/customer:
    post:
      consumes:
      - application/json
      description: Registers a customer
      parameters:
      - description: Customer informations
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/users.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Customer registered
          schema:
            type: string
        "409":
          description: Conflict
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Registers a customer
      tags:
      - authentication
  /auth/register/merchant:
    post:
      consumes:
      - application/json
      description: Registers a merchant
      parameters:
      - description: Merchant informations
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/users.RegisterMerchantRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Merchant registered
          schema:
            type: string
        "409":
          description: Conflict
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Registers a merchant
      tags:
      - authentication
  /auth/token:
    post:
      consumes:
      - application/json
      description: Creates a token for a user
      parameters:
      - description: User credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/users.CreateUserTokenRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Token
          schema:
            type: string
        "401":
          description: Unauthorized
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Creates a token
      tags:
      - authentication
  /checkout-sessions:
    post:
      consumes:
      - application/json
      description: Creates a checkout session
      parameters:
      - description: Checkout session request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/payments.CheckoutSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Checkout session created
          schema:
            type: string
        "403":
          description: Forbidden
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Creates a checkout session
      tags:
      - checkout-sessions
  /checkout-sessions/{token}/pay:
    post:
      description: Pays
      parameters:
      - description: Idempotency-Key must be set for valid response
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Checkout session token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Payment succeeded
          schema:
            type: string
        "402":
          description: Payment Required
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Pays
      tags:
      - checkout-sessions
  /customers/activate/{token}:
    put:
      description: Activates/Register a customer by invitation token
      parameters:
      - description: Invitation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Customer activated
          schema:
            type: string
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Activates/Register a customer
      tags:
      - users
  /merchants/activate/{token}:
    put:
      description: Activates/Register a merchant by invitation token
      parameters:
      - description: Invitation token
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Merchant activated
          schema:
            type: string
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Activates/Register a merchant
      tags:
      - users
  /refunds:
    post:
      description: Request a refund
      parameters:
      - description: Idempotency-Key must be set for valid response
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Refund request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/payments.RefundRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Refund accepted
          schema:
            type: string
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Request a refund
      tags:
      - refunds
  /wallet/top-ups:
    post:
      consumes:
      - application/json
      description: Initiates a topup
      parameters:
      - description: Idempotency-Key must be set for valid response
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Topup amount
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/wallets.TopupRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Topup initiated
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Initiates a topup
      tags:
      - wallet
  /wallet/transactions:
    get:
      description: Get user's transaction history
      produces:
      - application/json
      responses:
        "200":
          description: transactions
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Get user's transaction history
      tags:
      - wallet
  /wallet/transfers:
    post:
      consumes:
      - application/json
      description: P2P money transfer
      parameters:
      - description: Idempotency-Key must be set for valid response
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Transfer request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/wallets.TransferRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Transfer succeeded
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "402":
          description: Payment Required
          schema: {}
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "422":
          description: Unprocessable Entity
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: P2P money transfer
      tags:
      - wallet
  /wallet/withdrawals:
    post:
      consumes:
      - application/json
      description: Initiates a withdrawal
      parameters:
      - description: Idempotency-Key must be set for valid response
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Withdrawal amount
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/wallets.WithdrawalRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Withdrawal initiated
          schema:
            type: string
        "401":
          description: unauthorized
          schema:
            type: string
        "403":
          description: Forbidden
          schema: {}
        "404":
          description: Not Found
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      security:
      - ApiKeyAuth: []
      summary: Initiates a withdrawal
      tags:
      - wallet
  /webhooks/top-ups:
    post:
      consumes:
      - application/json
      description: Response from the PSP
      parameters:
      - description: X-Webhook-Signature must be set for valid response
        in: header
        name: X-Webhook-Signature
        required: true
        type: string
      - description: Webhook payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/wallets.TopupWebhookPayload'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema: {}
      summary: Response from the PSP
      tags:
      - webhooks
  /webhooks/withdrawals:
    post:
      consumes:
      - application/json
      description: Response from the PSP
      parameters:
      - description: X-Webhook-Signature must be set for valid response
        in: header
        name: X-Webhook-Signature
        required: true
        type: string
      - description: Webhook payload
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/wallets.PayoutWebhookPayload'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema: {}
      summary: Response from the PSP
      tags:
      - webhooks
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"